Rename to telemetry-operator
