+++ CODEOWNERS
@@ -253,8 +253,8 @@ components/etcd-tls-setup-job/ @PK85 @piotrmiskiewicz @ksputo @jasiu001 @adamwal
 # Central Connectivity Validator
 /components/central-application-connectivity-validator/ @akgalwas @janmedrek @franpog859 @Maladie @rafalpotempa @koala7659 @ralikio @everesio @ralikio @everesio
 
-# Telemetry Controller
-/components/telemetry-controller @a-thaler @lilitgh @suleymanakbas91 @clebs @rakesh-garimella @skhalash @jeremyharisch @chrkl
+# Telemetry Operator
+/components/telemetry-operator @a-thaler @lilitgh @suleymanakbas91 @clebs @rakesh-garimella @skhalash @jeremyharisch @chrkl
 
 # Tests
 # Service Catalog acceptance tests
