+++ components/telemetry-operator/.dockerignore
