+++ components/telemetry-operator/.gitignore
