+++ components/telemetry-operator/Dockerfile
