+++ components/telemetry-operator/PROJECT
@@ -1,8 +1,8 @@
 domain: kyma-project.io
 layout:
 - go.kubebuilder.io/v3
-projectName: telemetry-controller
-repo: github.com/kyma-project/kyma/components/telemetry-controller
+projectName: telemetry-operator
+repo: github.com/kyma-project/kyma/components/telemetry-operator
 resources:
 - api:
     crdVersion: v1
@@ -10,6 +10,6 @@ resources:
   domain: kyma-project.io
   group: telemetry
   kind: LoggingConfiguration
-  path: github.com/kyma-project/kyma/components/telemetry-controller/api/v1alpha1
+  path: github.com/kyma-project/kyma/components/telemetry-operator/api/v1alpha1
   version: v1alpha1
 version: "3"
