+++ components/telemetry-operator/api/v1alpha1/groupversion_info.go
