+++ components/telemetry-operator/api/v1alpha1/loggingconfiguration_types.go
