+++ components/telemetry-operator/config/crd/bases/telemetry.kyma-project.io_loggingconfigurations.yaml
