+++ components/telemetry-operator/config/crd/kustomization.yaml
