+++ components/telemetry-operator/config/crd/kustomizeconfig.yaml
