+++ components/telemetry-operator/config/crd/patches/cainjection_in_loggingconfigurations.yaml
