+++ components/telemetry-operator/config/crd/patches/webhook_in_loggingconfigurations.yaml
