+++ components/telemetry-operator/config/default/kustomization.yaml
@@ -6,7 +6,7 @@ namespace: kyma-system
 # "wordpress" becomes "alices-wordpress".
 # Note that it should also match with the prefix (text before '-') of the namespace
 # field above.
-namePrefix: telemetry-controller-
+namePrefix: telemetry-operator-
 
 # Labels to add to all resources and selectors.
 #commonLabels:
