+++ components/telemetry-operator/config/default/manager_config_patch.yaml
