+++ components/telemetry-operator/config/manager/controller_manager_config.yaml
