+++ components/telemetry-operator/config/manager/kustomization.yaml
