+++ components/telemetry-operator/config/manager/manager.yaml
