+++ components/telemetry-operator/config/prometheus/kustomization.yaml
