+++ components/telemetry-operator/config/prometheus/monitor.yaml
