+++ components/telemetry-operator/config/rbac/auth_proxy_role.yaml
