+++ components/telemetry-operator/config/rbac/auth_proxy_service.yaml
