+++ components/telemetry-operator/config/rbac/kustomization.yaml
