+++ components/telemetry-operator/config/rbac/leader_election_role_binding.yaml
