+++ components/telemetry-operator/config/rbac/loggingconfiguration_editor_role.yaml
