+++ components/telemetry-operator/config/rbac/loggingconfiguration_viewer_role.yaml
