+++ components/telemetry-operator/config/rbac/role.yaml
