+++ components/telemetry-operator/config/rbac/role_binding.yaml
