+++ components/telemetry-operator/config/rbac/service_account.yaml
