+++ components/telemetry-operator/config/samples/telemetry_v1alpha1_loggingconfiguration.yaml
