+++ components/telemetry-operator/hack/boilerplate.go.txt
